fix: resolve Runtime Fix 9 by extracting and integrating missing GeneralInfoStep component
